# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Kivro API
  version: 1.0.0
  description: |
    The Kivro API lets you generate, verify, and track digital addresses 
    across all 54 African countries. Use `/api/v1/` for all new integrations.
  contact:
    name: Kivro Developer Support
    url: https://kivro.africa/help
    email: developers@kivro.africa
  license:
    name: Proprietary
    url: https://kivro.africa/terms
servers:
  - url: https://kivro-fullstack.vercel.app/api/v1
    description: Production
  - url: http://localhost:3001/api/v1
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Pass your Kivro API key as `Bearer kv_live_<key>` 
        OR your Supabase JWT for user-authenticated endpoints.
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:               { type: string, format: uuid }
        name:             { type: string }
        key_prefix:       { type: string, example: "kv_live_ab12cd34" }
        plan:             { type: string, enum: [pro, business, enterprise, government] }
        status:           { type: string, enum: [active, revoked] }
        rate_limit_per_min: { type: integer }
        requests_today:   { type: integer }
        created_at:       { type: string, format: date-time }
...
paths:
...
  /keys:
    get:
      summary: List API keys
      tags: [API Keys]
      responses:
        '200':
          description: List of keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys: { type: array, items: { $ref: '#/components/schemas/ApiKey' } }
    post:
      summary: Create an API key
      tags: [API Keys]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, plan]
              properties:
                name: { type: string }
                plan: { type: string, enum: [pro, business, enterprise, government] }
      responses:
        '201':
          description: Key created — plaintext shown once
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:       { $ref: '#/components/schemas/ApiKey' }
                  plaintext: { type: string, example: "kv_live_ab12cd34..." }
  /keys/{id}:
    delete:
      summary: Revoke an API key
      tags: [API Keys]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Key revoked }
        '404': { description: Key not found }
...
  /billing/subscribe:
    post:
      summary: Create or upgrade subscription
      tags: [Billing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan:     { type: string, enum: [free, pro, business, enterprise, government] }
                currency: { type: string, example: XOF }
                interval: { type: string, enum: [month, year] }
      responses:
        '201': { description: Subscription created }
...
  /billing/revenue/{period}:
    get:
      summary: Revenue summary for a billing period
      tags: [Billing]
      parameters:
        - name: period
          in: path
          required: true
          schema: { type: string, example: '2026-05' }
      responses:
        '200':
          description: Revenue breakdown
...
  /packages:
    get:
      summary: List packages (admin/courier)
      tags: [Packages]
      responses:
        '200': { description: Package list }
    post:
      summary: Create a package
      tags: [Packages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sender_name, recipient_name, recipient_address, qr_code]
              properties:
                sender_name:       { type: string }
                recipient_name:    { type: string }
                recipient_address: { type: string }
                qr_code:           { type: string }
      responses:
        '201': { description: Package created }
  /packages/{id}/status:
    patch:
      summary: Update package status
      tags: [Packages]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [pending, picked_up, in_transit, delivered, failed] }
                notes:  { type: string }
      responses:
        '200': { description: Status updated }
...